# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/programs": {
      "get": {
        "description": "Public — returns all active programs",
        "operationId": "ProgramsController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Programs retrieved successfully"
          }
        },
        "summary": "Get all programs",
        "tags": [
          "Programs"
        ]
      },
      "post": {
        "description": "Org admins and superusers can create programs",
        "operationId": "ProgramsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgramDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Program created successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "403": {
            "description": "Access denied — requires admin or superuser role"
          }
        },
        "summary": "Create a new program",
        "tags": [
          "Programs"
        ]
      }
    },
    "/programs/by-org/{org}": {
      "get": {
        "operationId": "ProgramsController_findByOrg",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "org",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Programs retrieved"
          }
        },
        "summary": "Get programs by organization name",
        "tags": [
          "Programs"
        ]
      }
    },
    "/programs/{id}": {
      "get": {
        "operationId": "ProgramsController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program retrieved"
          },
          "404": {
            "description": "Program not found"
          }
        },
        "summary": "Get program by ID",
        "tags": [
          "Programs"
        ]
      },
      "patch": {
        "operationId": "ProgramsController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgramDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Program updated"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Program not found"
          }
        },
        "summary": "Update a program",
        "tags": [
          "Programs"
        ]
      },
      "delete": {
        "operationId": "ProgramsController_remove",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program deleted"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Program not found"
          }
        },
        "summary": "Delete a program",
        "tags": [
          "Programs"
        ]
      }
    },
...
    "/program-assignments": {
      "get": {
        "operationId": "ProgramAssignmentsController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments retrieved"
          },
          "403": {
            "description": "Access denied — requires admin or superuser role"
          }
        },
        "summary": "Get all program assignments",
        "tags": [
          "Program Assignments"
        ]
      },
      "post": {
        "description": "Assign a volunteer to a program",
        "operationId": "ProgramAssignmentsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgramAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assignment created successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Create a program assignment",
        "tags": [
          "Program Assignments"
        ]
      }
    },
    "/program-assignments/by-program/{programId}": {
      "get": {
        "operationId": "ProgramAssignmentsController_findByProgram",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get assignments by program ID",
        "tags": [
          "Program Assignments"
        ]
      }
    },
    "/program-assignments/by-volunteer/{email}": {
      "get": {
        "operationId": "ProgramAssignmentsController_findByVolunteer",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get assignments by volunteer email",
        "tags": [
          "Program Assignments"
        ]
      }
    },
    "/program-assignments/{id}": {
      "get": {
        "operationId": "ProgramAssignmentsController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment retrieved"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Assignment not found"
          }
        },
        "summary": "Get assignment by ID",
        "tags": [
          "Program Assignments"
        ]
      },
      "patch": {
        "operationId": "ProgramAssignmentsController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgramAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignment updated"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Assignment not found"
          }
        },
        "summary": "Update a program assignment",
        "tags": [
          "Program Assignments"
        ]
      },
      "delete": {
        "operationId": "ProgramAssignmentsController_remove",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment deleted"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Assignment not found"
          }
        },
        "summary": "Delete a program assignment",
        "tags": [
          "Program Assignments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kindred Platform API",
    "description": "Backend REST API for Kindred — India's Volunteer Platform. Supports Users, Organizations, Volunteers, Programs, Requests, Tasks, Hour Logs, Resource Donations, Program Assignments, Program Applications, and Subtasks. Role-based access control via the \"role\" header.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateProgramDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Program name",
            "example": "Rural Literacy Drive 2025"
          },
          "org": {
            "type": "string",
            "description": "Organizing NGO",
            "example": "GreenFuture Foundation"
          },
          "focus": {
            "type": "string",
            "description": "Focus area",
            "example": "Education"
          },
          "startDate": {
            "type": "string",
            "description": "Start date (YYYY-MM-DD)",
            "example": "2025-02-01"
          },
          "endDate": {
            "type": "string",
            "description": "End date (YYYY-MM-DD)",
            "example": "2025-08-31"
          },
          "budget": {
            "type": "number",
            "description": "Total budget in INR",
            "example": 850000
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the program",
            "example": "Providing free education materials and tutoring to 500 children in remote villages of Jharkhand"
          },
          "requiredSkills": {
            "description": "Skills needed from volunteers",
            "example": [
              "Teaching",
              "Hindi Translation",
              "Logistics"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "volunteerGoal": {
            "type": "number",
            "description": "Target number of volunteers",
            "example": 80
          },
          "openForApplications": {
            "type": "boolean",
            "description": "Whether volunteers can apply",
            "example": true
          },
          "location": {
            "type": "string",
            "description": "Geographic location",
            "example": "Ranchi, Jharkhand"
          }
        },
        "required": [
          "name",
          "org",
          "focus",
          "startDate",
          "endDate",
          "budget",
          "description"
        ]
      },
      "UpdateProgramDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Program name",
            "example": "Rural Literacy Drive 2025"
          },
          "org": {
            "type": "string",
            "description": "Organizing NGO",
            "example": "GreenFuture Foundation"
          },
          "focus": {
            "type": "string",
            "description": "Focus area",
            "example": "Education"
          },
          "startDate": {
            "type": "string",
            "description": "Start date (YYYY-MM-DD)",
            "example": "2025-02-01"
          },
          "endDate": {
            "type": "string",
            "description": "End date (YYYY-MM-DD)",
            "example": "2025-08-31"
          },
          "budget": {
            "type": "number",
            "description": "Total budget in INR",
            "example": 850000
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the program",
            "example": "Providing free education materials and tutoring to 500 children in remote villages of Jharkhand"
          },
          "requiredSkills": {
            "description": "Skills needed from volunteers",
            "example": [
              "Teaching",
              "Hindi Translation",
              "Logistics"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "volunteerGoal": {
            "type": "number",
            "description": "Target number of volunteers",
            "example": 80
          },
          "openForApplications": {
            "type": "boolean",
            "description": "Whether volunteers can apply",
            "example": true
          },
          "location": {
            "type": "string",
            "description": "Geographic location",
            "example": "Ranchi, Jharkhand"
          },
          "status": {
            "type": "string",
            "example": "active"
          },
          "raised": {
            "type": "number",
            "example": 850000
          },
          "beneficiaries": {
            "type": "number",
            "example": 1500
          },
          "volunteers": {
            "type": "number",
            "example": 120
          }
        }
      },
...
      "CreateProgramAssignmentDto": {
        "type": "object",
        "properties": {
          "programId": {
            "type": "string",
            "description": "Program ID",
            "example": "prog_001"
          },
          "programName": {
            "type": "string",
            "description": "Program name",
            "example": "Cyclone Relief Fund"
          },
          "volunteerEmail": {
            "type": "string",
            "description": "Volunteer email",
            "example": "neha.kapoor@gmail.com"
          },
          "volunteerName": {
            "type": "string",
            "description": "Volunteer name",
            "example": "Neha Kapoor"
          },
          "role": {
            "type": "string",
            "description": "Role in the program",
            "example": "Field Coordinator"
          }
        },
        "required": [
          "programId",
          "programName",
          "volunteerEmail",
          "volunteerName"
        ]
      },
      "UpdateProgramAssignmentDto": {
        "type": "object",
        "properties": {
          "programId": {
            "type": "string",
            "description": "Program ID",
            "example": "prog_001"
          },
          "programName": {
            "type": "string",
            "description": "Program name",
            "example": "Cyclone Relief Fund"
          },
          "volunteerEmail": {
            "type": "string",
            "description": "Volunteer email",
            "example": "neha.kapoor@gmail.com"
          },
          "volunteerName": {
            "type": "string",
            "description": "Volunteer name",
            "example": "Neha Kapoor"
          },
          "role": {
            "type": "string",
            "description": "Role in the program",
            "example": "Field Coordinator"
          }
        }
      },
...
    }
  }
}